+++ CHANGES.md
@@ -16,6 +16,7 @@
 * Update Six compatibility library 1.13.0 (ec58185) to 1.14.0 (3a3db75)
 * Update Tornado_py3 Web Server 6.0.3 (ff985fe) to 6.0.xx (18b653c)
 * Update urllib3 release 1.25.6 (4a6c288) to 1.25.7 (37ba61a)
+* Add Telegram notifier
 
 
 [develop changelog]